python/requirements.txt
pyannote.audio>=3.1
torch>=2.0
torchaudio>=2.0
